Data.List.Clojure.Generic

Description

Generic variants of Clojure-inspired list functions for Haskell.

Documentation

dropLast1 ∷ [a] → [a]

Return all but the last element of a given list.

dropLast ∷ Integral b ⇒ b → [a] → [a]

Given a number n and a list l, return a list of all but the last n elements in l.

takeNth ∷ Integral b ⇒ b → [a] → [a]

Return a list of every nth element of a given list.